reorganize repo file structure and exports
